moves routes to api dir
