+++ server/routes/api/auth.go
@@ -1,4 +1,4 @@
-package routes
+package api
 
 import (
 	"github.com/go-chi/chi"
