+++ server/routes/api/products.go
@@ -1,10 +1,10 @@
-package routes
+package api
 
 import (
 	"github.com/go-chi/chi"
 )
 
-func NewStoreRoutes() chi.Router {
+func NewProductsRoutes() chi.Router {
 	r := chi.NewRouter()
 
 	return r
