+++ server/routes/api/stores.go
@@ -1,10 +1,10 @@
-package routes
+package api
 
 import (
 	"github.com/go-chi/chi"
 )
 
-func NewProductsRoutes() chi.Router {
+func NewStoreRoutes() chi.Router {
 	r := chi.NewRouter()
 
 	return r
